requirements.txt
requests
pyzipper
python-dotenv
